requirements.txt
streamlit
pandas
qrcode
Pillow
openpyxl
xlsxwriter
